fix(reef): split legacy readers to resolve max-lines and curly lint
